Updated source
Clicking on a particular section of the screen does not work so well
when you are dealing with smaller resolutions. The function has since
been re-written to do a pixel search on the client window to look for
certain colors. This has been tested on smaller resolutions and seems to
work.