Updated remember username color
With the login screen change the color code for detecting if the
checkbox was checked has changed. This might need to be changed with
each patch. I might throw the color code up in a text file in the future
so I dont have to keep creating a new version for each patch.