Fixed Gnar bug
If there was a similar color on the screen as the login button AKA Gnar
the script would click on that portion of the screen if there was a
patch. The bounding box has been changed so the script only checks the
top portion of the screen. Since this change was made I have decreased
the sensitivity of the pixel search so that the button can be found
easier.