New riot patch
Riot changed the location of the league client window information in the
WinGetPos array list so the location was updated.

The remember username checkbox now functions as it should which puts the
focus in the password field by default. A check was added to see if that
checkbox was checked or not.