Monitor fixes
Moved the check window function to a new function since it was being
called so much. There is probably a better way to do this but I am not
sure of it at the moment. Also due to the fact that the username and
password field load a  lot sooner there is no need for any more sleep
delays.